# swagger.json
{"openapi":"3.0.0","paths":{"/healthcheck":{"get":{"operationId":"HealthcheckController_getStatus","summary":"Get server status","parameters":[],"responses":{"200":{"description":"ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthcheckDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["healthcheck"]}},"/customer/info":{"get":{"operationId":"CustomerController_getCustomer","summary":"Get the customer data","parameters":[],"responses":{"200":{"description":"Returns a JSON with the customer data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetCustomerResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["customer"],"security":[{"auth":[]}]}},"/customer/insert":{"put":{"operationId":"CustomerController_putCustomer","summary":"Put the customer data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutCustomerReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the customer data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetCustomerResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["customer"],"security":[{"auth":[]}]}},"/customer/remove":{"delete":{"operationId":"CustomerController_deleteCustomer","summary":"Delete the customer data","parameters":[],"responses":{"200":{"description":"Returns a JSON with the customer status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteCustomerResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["customer"],"security":[{"auth":[]}]}}},"info":{"title":"pdp-ms-customer","description":"cosmetics e-commerce","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"/pdp-ms-customer"}],"components":{"securitySchemes":{"auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","in":"header"}},"schemas":{"HealthcheckDto":{"type":"object","properties":{"name":{"type":"string"},"version":{"type":"string"}},"required":["name","version"]},"ErrorDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"CustomerDto":{"type":"object","properties":{"cpf":{"type":"string"},"phone":{"type":"string"},"zipcode":{"type":"string"},"city":{"type":"string"},"complement":{"type":"string"},"street":{"type":"string"},"number":{"type":"number"},"neighbourhood":{"type":"string"},"state":{"type":"string"}},"required":["cpf","phone","zipcode","city","complement","street","number","neighbourhood","state"]},"GetCustomerResDto":{"type":"object","properties":{"customer":{"$ref":"#/components/schemas/CustomerDto"}},"required":["customer"]},"PutCustomerReqDto":{"type":"object","properties":{"cpf":{"type":"string"},"phone":{"type":"string"},"zipcode":{"type":"string"},"city":{"type":"string"},"complement":{"type":"string"},"street":{"type":"string"},"number":{"type":"number"},"neighbourhood":{"type":"string"},"state":{"type":"string"}},"required":["cpf","phone","zipcode","city","complement","street","number","neighbourhood","state"]},"DeleteCustomerResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]}}}}